split coral-mcp surface helpers
